+++ CHANGES.rst
@@ -1,6 +1,12 @@
 Change log
 ==========
 
+v3.2.6
+------
+* Extensive support for elasticity tensor analysis (Joseph Montoya).
+* Misc bug fixes and performance improvements.
+* Add support for QChem4.3 new format of Batch jobs
+
 v3.2.5
 ------
 * Improved potcar setup via "pmg setup", with MAPI setup.